Lifted userDetail to ContactCard state
